Add jwt

+++ src/controllers/employee.ts
@@ -2,6 +2,7 @@ import { RequestHandler } from "express";
 import { isEqual } from "lodash";
 import { employee, employeeSchema } from "../models/employee";
 import { db } from "../models";
+import jwt from "jsonwebtoken";
 
 const EMPLOYEES = db.employees;
 
@@ -10,7 +11,18 @@ const idFormatError = (res: any) =>
     errorMessage: "The ID format is wrong!",
   });
 
+const verifyToken = (req: any, res: any) => {
+  const token = req.headers["x-access-token"];
+  try {
+    jwt.verify(token as string, process.env.SECRET_KEY as string);
+    return true;
+  } catch (err) {
+    return false;
+  }
+};
+
 export const createEmp: RequestHandler = (req, res) => {
+  if (!verifyToken(req, res)) return res.status(500).send("Need log in.");
   const input = req.body as employee;
   const { value, error } = employeeSchema.validate(input);
   if (error) return res.status(400).json({ errorMessage: error.message });
@@ -20,6 +32,7 @@ export const createEmp: RequestHandler = (req, res) => {
 };
 
 export const getEmp: RequestHandler = (req, res) => {
+  if (!verifyToken(req, res)) return res.status(500).send("Need log in.");
   if (req.params.id && isNaN(+req.params.id)) return idFormatError(res);
   EMPLOYEES.findAll({
     order: [["id", "asc"]],
@@ -34,6 +47,7 @@ export const getEmp: RequestHandler = (req, res) => {
 };
 
 export const updateEmp: RequestHandler = (req, res) => {
+  if (!verifyToken(req, res)) return res.status(500).send("Need log in.");
   const input = req.body as employee;
   if (isNaN(+req.params.id)) return idFormatError(res);
   EMPLOYEES.findByPk(+req.params.id).then((data) => {
@@ -61,6 +75,7 @@ export const updateEmp: RequestHandler = (req, res) => {
 };
 
 export const delEmp: RequestHandler = (req, res) => {
+  if (!verifyToken(req, res)) return res.status(500).send("Need log in.");
   if (isNaN(+req.params.id)) return idFormatError(res);
   EMPLOYEES.destroy({
     where: { id: +req.params.id },
